v0.5.0-beta: centralised menu
