+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.5.0-beta]
+### Changed
+- Refactored launch logic to remove WET code and have a centralised menu
+
+
 ## [v0.4.5] - 2025-11-24
 ### Added
 - Simple directory browser for choosing which directory to compress with Windows Explorer for better compatibility with Windows' varying flavours of CLI terminals
