+++ locales/de.json
@@ -61,13 +61,7 @@
     "LZX compression requires at least {min_physical} cores and {min_logical} threads.": "LZX-Komprimierung erfordert mindestens {min_physical} Kerne und {min_logical} Threads.",
     "LZX compression has been disabled for better system responsiveness.": "LZX-Komprimierung wurde für eine bessere Systemreaktionsfähigkeit deaktiviert.",
     "Use -f flag to force LZX if you're feeling adventurous.": "Verwenden Sie das Flag -f, um LZX zu erzwingen, wenn Sie abenteuerlustig sind.",
-    "Enter directory path to compress: ": "Geben Sie den Verzeichnispfad zum Komprimieren ein: ",
-    "Directory path cannot be empty.": "Verzeichnispfad darf nicht leer sein.",
-    "Unable to resolve the target volume. Please verify the path.": "Zielvolume konnte nicht aufgelöst werden. Bitte überprüfen Sie den Pfad.",
-    "Network shares are not supported targets for compression.": "Netzwerkfreigaben sind keine unterstützten Ziele für die Komprimierung.",
-    "Please select a local NTFS volume instead.": "Bitte wählen Sie stattdessen ein lokales NTFS-Volume.",
-    "Windows compression requires NTFS.": "Windows-Komprimierung erfordert NTFS.",
-    "Detected filesystem: {filesystem}": "Erkanntes Dateisystem: {filesystem}",
+
     "Detected a traditional spinning hard drive for this path.": "Traditionelle Festplatte (HDD) für diesen Pfad erkannt.",
     "Sustained compression can thrash the disk heads, fragment files, and slow app/game launches.": "Anhaltende Komprimierung kann die Schreib-/Leseköpfe belasten, Dateien fragmentieren und App-/Spielstarts verlangsamen.",
     "\nRecommendation:": "\nEmpfehlung:",
