+++ locales/es.json
@@ -61,13 +61,7 @@
     "LZX compression requires at least {min_physical} cores and {min_logical} threads.": "La compresión LZX requiere al menos {min_physical} núcleos y {min_logical} hilos.",
     "LZX compression has been disabled for better system responsiveness.": "La compresión LZX ha sido desactivada para mejor respuesta del sistema.",
     "Use -f flag to force LZX if you're feeling adventurous.": "Use la bandera -f para forzar LZX si se siente aventurero.",
-    "Enter directory path to compress: ": "Ingrese la ruta del directorio a comprimir: ",
-    "Directory path cannot be empty.": "La ruta del directorio no puede estar vacía.",
-    "Unable to resolve the target volume. Please verify the path.": "No se puede resolver el volumen objetivo. Por favor verifique la ruta.",
-    "Network shares are not supported targets for compression.": "Los recursos compartidos de red no son objetivos soportados para compresión.",
-    "Please select a local NTFS volume instead.": "Por favor seleccione un volumen NTFS local en su lugar.",
-    "Windows compression requires NTFS.": "La compresión de Windows requiere NTFS.",
-    "Detected filesystem: {filesystem}": "Sistema de archivos detectado: {filesystem}",
+
     "Detected a traditional spinning hard drive for this path.": "Detectado un disco duro tradicional (HDD) para esta ruta.",
     "Sustained compression can thrash the disk heads, fragment files, and slow app/game launches.": "La compresión sostenida puede agotar los cabezales del disco, fragmentar archivos y ralentizar lanzamientos de apps/juegos.",
     "\nRecommendation:": "\nRecomendación:",
