+++ locales/fr.json
@@ -61,13 +61,7 @@
     "LZX compression requires at least {min_physical} cores and {min_logical} threads.": "La compression LZX nécessite au moins {min_physical} cœurs et {min_logical} threads.",
     "LZX compression has been disabled for better system responsiveness.": "La compression LZX a été désactivée pour une meilleure réactivité du système.",
     "Use -f flag to force LZX if you're feeling adventurous.": "Utilisez le drapeau -f pour forcer LZX si vous vous sentez aventureux.",
-    "Enter directory path to compress: ": "Entrez le chemin du répertoire à compresser : ",
-    "Directory path cannot be empty.": "Le chemin du répertoire ne peut pas être vide.",
-    "Unable to resolve the target volume. Please verify the path.": "Impossible de résoudre le volume cible. Veuillez vérifier le chemin.",
-    "Network shares are not supported targets for compression.": "Les partages réseau ne sont pas des cibles supportées pour la compression.",
-    "Please select a local NTFS volume instead.": "Veuillez sélectionner un volume NTFS local à la place.",
-    "Windows compression requires NTFS.": "La compression Windows nécessite NTFS.",
-    "Detected filesystem: {filesystem}": "Système de fichiers détecté : {filesystem}",
+
     "Detected a traditional spinning hard drive for this path.": "Disque dur traditionnel (HDD) détecté pour ce chemin.",
     "Sustained compression can thrash the disk heads, fragment files, and slow app/game launches.": "Une compression soutenue peut solliciter excessivement les têtes de lecture, fragmenter les fichiers et ralentir le lancement des applications/jeux.",
     "\nRecommendation:": "\nRecommandation :",
