+++ locales/pt.json
@@ -61,13 +61,7 @@
     "LZX compression requires at least {min_physical} cores and {min_logical} threads.": "A compressão LZX requer pelo menos {min_physical} núcleos e {min_logical} threads.",
     "LZX compression has been disabled for better system responsiveness.": "A compressão LZX foi desativada para melhor capacidade de resposta do sistema.",
     "Use -f flag to force LZX if you're feeling adventurous.": "Use a flag -f para forçar LZX se estiver se sentindo aventureiro.",
-    "Enter directory path to compress: ": "Insira o caminho do diretório para comprimir: ",
-    "Directory path cannot be empty.": "O caminho do diretório não pode estar vazio.",
-    "Unable to resolve the target volume. Please verify the path.": "Não foi possível resolver o volume de destino. Por favor, verifique o caminho.",
-    "Network shares are not supported targets for compression.": "Compartilhamentos de rede não são alvos suportados para compressão.",
-    "Please select a local NTFS volume instead.": "Por favor, selecione um volume NTFS local em vez disso.",
-    "Windows compression requires NTFS.": "A compressão do Windows requer NTFS.",
-    "Detected filesystem: {filesystem}": "Sistema de arquivos detectado: {filesystem}",
+
     "Detected a traditional spinning hard drive for this path.": "Detectado um disco rígido tradicional (HDD) para este caminho.",
     "Sustained compression can thrash the disk heads, fragment files, and slow app/game launches.": "A compressão sustentada pode desgastar as cabeças do disco, fragmentar arquivos e tornar o lançamento de apps/jogos lento.",
     "\nRecommendation:": "\nRecomendação:",
